% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_calculator.R
\name{age_calculator}
\alias{age_calculator}
\title{Calculate your age from your birthyear.}
\usage{
age_calculator(birthyear)
}
\arguments{
\item{birthyear}{Numeric value or vector.}
}
\value{
Numeric value or vector.
}
\description{
Calculate your age from your birthyear.
}
\examples{
age_calculator(2000)
age_calculator(c(1985, 2000))
}
