% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_age.R
\name{current_age}
\alias{current_age}
\title{My current age function.}
\usage{
current_age(birthdate)
}
\arguments{
\item{birthdate}{A birthdate in the format of "YYYY-MM-DD".}
}
\value{
The current age of someone with the given birthdate.
}
\description{
My current age function.
}
\examples{
current_age("1967-06-30")

birthdate <- "1967-06-30"
current_age(birthdate)
}
