% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_year.R
\name{current_year}
\alias{current_year}
\title{Current Year Function}
\usage{
current_year(year)
}
\arguments{
\item{year}{Generated current year using the `sys.date()` command formatted to a four-digit long year.}
}
\value{
A character string stating "The year is `year`"
}
\description{
Current Year Function
}
\examples{
current_year(year)
}
