% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-calculation.R
\name{my_age}
\alias{my_age}
\title{Calculate your age for the current year}
\usage{
my_age(birth_year)
}
\arguments{
\item{birth_year}{enter your birth year as a numerical value}
}
\value{
calculates your age by subtraction of current year with your birth year
}
\description{
Calculate your age for the current year
}
