% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fav-color-function.R
\name{my_favorite_color}
\alias{my_favorite_color}
\title{Print out your favorite color}
\usage{
my_favorite_color(x)
}
\arguments{
\item{x}{x will be your favorite color}
}
\value{
pastes out "___ is my favorite color"
}
\description{
Print out your favorite color
}
