% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_name.R
\name{my_name}
\alias{my_name}
\title{My name and favorite color function.}
\usage{
my_name(name, color)
}
\arguments{
\item{name}{Character value of a name.}

\item{color}{Character value of a color.}
}
\value{
A sentence stating a person's (my) name and their (my) favorite color.
}
\description{
My name and favorite color function.
}
\examples{
my_name("Shayna","blue")
}
