% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_color.R
\name{name_color}
\alias{name_color}
\title{name_color}
\usage{
name_color(name)
}
\arguments{
\item{name}{A character string, preferably a name.}
}
\value{
A character string that reports your name and your favorite color.
}
\description{
name_color
}
\examples{
name_color(name = "Brady")
name_color(name = "Braden")
name_color(name = "B.C.")
name_color(name = "Jim")
}
