% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_name_color.R
\name{real_name_color}
\alias{real_name_color}
\title{real_name_color}
\usage{
real_name_color(name, color)
}
\arguments{
\item{name}{A character string.}

\item{color}{A character string.}
}
\value{
A character string, a longer one.
}
\description{
real_name_color
}
\examples{
real_name_color(name = "Brady", color = "chartreuse")
}
