% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/your_age.R
\name{your_age}
\alias{your_age}
\title{your_age}
\usage{
your_age(birthyr)
}
\arguments{
\item{birthyr}{A numeric value, preferably your birth year in the YYYY format.}
}
\value{
A numeric value.
}
\description{
your_age
}
\examples{
your_age(birthyr = 1995)
}
