% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name-function.R
\name{my_name}
\alias{my_name}
\title{Function that prints out your name}
\usage{
my_name(x)
}
\arguments{
\item{x}{x would be your name}
}
\value{
pastes out "__ is my name"
}
\description{
Function that prints out your name
}
