% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{name_color}
\alias{name_color}
\title{Print Name Function}
\usage{
name_color(name, color)
}
\arguments{
\item{name}{Character string of my name}
}
\value{
Character string stating "My name is {name}"
}
\description{
Print Name Function
}
\examples{
name_color("Craig","blue")
}
