% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_color.R
\name{name_color}
\alias{name_color}
\title{Announce your name and favorite color}
\usage{
name_color(your_name, fav_color)
}
\arguments{
\item{your_name}{Character value or vector.}

\item{fav_color}{Character value or vector.}
}
\value{
Character value or vector.
}
\description{
Announce your name and favorite color
}
\examples{
name_color("Cookie Monster", "chocolate brown")
name_color(c("Cookie Monster", "Ariel"), c("chocolate brown","ocean blue"))
}
